fix(mcp): 전역 서버 /start·/stop 가 유저풀 대신 전역 registry 로 — 소유자 불일치 500 해소
POST /api/mcp/servers/:id/start 가 전역(visibility=global) 서버도 spawnUserServer(actor.id)
로 보내 supervisor 가 `서버 소유자 불일치` 로 throw → 500 + 전역 서버에 actor 명의
crashed 이력이 남았다 (2026-08-25 noapi-google-search 실측). /connect 의 전역 분기
(registry 연결 + env 복호화)를 mcp-global-connect.ts 로 뽑아 /connect·/start 가 공유하고,
/stop 도 전역이면 registry.disconnectServer 로 대칭 처리.
